# docs/swagger.yaml
definitions:
...
  server.AssignPermissionRequest:
    type: object
  server.AssignPermissionResponse:
    type: object
...
  server.CreatePermissionRequest:
    type: object
  server.CreatePermissionResponse:
    type: object
...
  server.CreateReBACSchemaRequest:
    type: object
  server.CreateReBACSchemaResponse:
    type: object
  server.CreateRelationTypeRequest:
    type: object
  server.CreateRelationTypeResponse:
    type: object
  server.CreateRoleRequest:
    properties:
      description:
        description: |-
          Description of the role
          example: Administrator role with full access
        type: string
      name:
        description: |-
          The name of the role
          required: true
          example: admin
        type: string
      properties:
        additionalProperties:
          type: string
        description: |-
          Additional properties for the role
          example: {"department": "IT", "level": "high"}
        type: object
    required:
    - name
    type: object
  server.CreateRoleResponse:
    properties:
      message:
        description: |-
          Success message
          example: Role created successfully
        type: string
      role:
        allOf:
        - $ref: '#/definitions/server.RoleDetails'
        description: Created role details
    type: object
...
  server.ErrorResponse:
    properties:
      error:
        description: |-
          Error code
          example: validation_error
        type: string
      message:
        description: |-
          Error message
          example: Role name is required
        type: string
    type: object
...
  server.PermissionDetails:
    type: object
...
  server.ReBACSchemaDetails:
    type: object
  server.RelationTypeDetails:
    type: object
  server.RemovePermissionRequest:
    type: object
  server.RemovePermissionResponse:
    type: object
  server.RoleDetails:
    properties:
      description:
        description: |-
          Description of the role
          example: Administrator role with full access
        type: string
      entity_id:
        description: |-
          Unique entity identifier
          example: role_admin_123abc
        type: string
      name:
        description: |-
          The name of the role
          example: admin
        type: string
      properties:
        additionalProperties:
          type: string
        description: |-
          Additional properties for the role
          example: {"department": "IT", "level": "high", "type": "role", "parent": "role"}
        type: object
    type: object
...
  server.UpdateReBACSchemaRequest:
    type: object
  server.UpdateReBACSchemaResponse:
    type: object
  server.UpdateRoleRequest:
    type: object
  server.UpdateRoleResponse:
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/rbac/permissions:
    get:
      description: Get a list of all permissions in the RBAC system
      produces:
      - application/json
      responses:
        "200":
          description: List of permissions
          schema:
            items:
              $ref: '#/definitions/server.PermissionDetails'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: List all permissions
      tags:
      - 2-advanced-rbac
    post:
      consumes:
      - application/json
      description: Create a new permission in the RBAC system
      parameters:
      - description: Permission creation request
        in: body
        name: permission
        required: true
        schema:
          $ref: '#/definitions/server.CreatePermissionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Permission created successfully
          schema:
            $ref: '#/definitions/server.CreatePermissionResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: Create a new permission
      tags:
      - 2-advanced-rbac
  /api/v1/rbac/permissions/assign:
    post:
      consumes:
      - application/json
      description: Assign an existing permission to a role in the RBAC system
      parameters:
      - description: Permission assignment request
        in: body
        name: assignment
        required: true
        schema:
          $ref: '#/definitions/server.AssignPermissionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Permission assigned successfully
          schema:
            $ref: '#/definitions/server.AssignPermissionResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "404":
          description: Role or permission not found
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: Assign permission to role
      tags:
      - 2-advanced-rbac
  /api/v1/rbac/permissions/remove:
    post:
      consumes:
      - application/json
      description: Remove an assigned permission from a role in the RBAC system
      parameters:
      - description: Permission removal request
        in: body
        name: removal
        required: true
        schema:
          $ref: '#/definitions/server.RemovePermissionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Permission removed successfully
          schema:
            $ref: '#/definitions/server.RemovePermissionResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "404":
          description: Role or permission not found
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: Remove permission from role
      tags:
      - 2-advanced-rbac
  /api/v1/rbac/roles:
    get:
      description: Get a list of all roles in the RBAC system
      produces:
      - application/json
      responses:
        "200":
          description: List of roles
          schema:
            items:
              $ref: '#/definitions/server.RoleDetails'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: List all roles
      tags:
      - 2-advanced-rbac
    post:
      consumes:
      - application/json
      description: Create a new role in the RBAC system as a subject attribute
      parameters:
      - description: Role creation request
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/server.CreateRoleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Role created successfully
          schema:
            $ref: '#/definitions/server.CreateRoleResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: Create a new role
      tags:
      - 2-advanced-rbac
  /api/v1/rbac/roles/{id}:
    delete:
      description: Delete an existing role from the RBAC system
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: Role deleted successfully
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "404":
          description: Role not found
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: Delete a role
      tags:
      - 2-advanced-rbac
    put:
      consumes:
      - application/json
      description: Update an existing role in the RBAC system
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      - description: Role update request
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/server.UpdateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Role updated successfully
          schema:
            $ref: '#/definitions/server.UpdateRoleResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "404":
          description: Role not found
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: Update a role
      tags:
      - 2-advanced-rbac
  /api/v1/rebac/relation-types:
    get:
      description: Get a list of all relation types used in ReBAC schemas
      produces:
      - application/json
      responses:
        "200":
          description: List of relation types
          schema:
            items:
              $ref: '#/definitions/server.RelationTypeDetails'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: List all relation types
      tags:
      - 4-advanced-rebac
    post:
      consumes:
      - application/json
      description: Create a new relation type for ReBAC schemas
      parameters:
      - description: Relation type creation request
        in: body
        name: relationType
        required: true
        schema:
          $ref: '#/definitions/server.CreateRelationTypeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Relation type created successfully
          schema:
            $ref: '#/definitions/server.CreateRelationTypeResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: Create a new relation type
      tags:
      - 4-advanced-rebac
  /api/v1/rebac/schemas:
    get:
      description: Get a list of all Relationship-Based Access Control schemas
      produces:
      - application/json
      responses:
        "200":
          description: List of ReBAC schemas
          schema:
            items:
              $ref: '#/definitions/server.ReBACSchemaDetails'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: List all ReBAC schemas
      tags:
      - 4-advanced-rebac
    post:
      consumes:
      - application/json
      description: Create a new Relationship-Based Access Control schema
      parameters:
      - description: ReBAC schema creation request
        in: body
        name: schema
        required: true
        schema:
          $ref: '#/definitions/server.CreateReBACSchemaRequest'
      produces:
      - application/json
      responses:
        "201":
          description: ReBAC schema created successfully
          schema:
            $ref: '#/definitions/server.CreateReBACSchemaResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: Create a new ReBAC schema
      tags:
      - 4-advanced-rebac
  /api/v1/rebac/schemas/{id}:
    put:
      consumes:
      - application/json
      description: Update an existing Relationship-Based Access Control schema
      parameters:
      - description: Schema ID
        in: path
        name: id
        required: true
        type: string
      - description: ReBAC schema update request
        in: body
        name: schema
        required: true
        schema:
          $ref: '#/definitions/server.UpdateReBACSchemaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ReBAC schema updated successfully
          schema:
            $ref: '#/definitions/server.UpdateReBACSchemaResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "404":
          description: Schema not found
          schema:
            $ref: '#/definitions/server.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: Update a ReBAC schema
      tags:
      - 4-advanced-rebac
...
swagger: "2.0"