# pkg/server/rbac_handlers.go
package server

import (
	"encoding/json"
	"net/http"
)

// RBAC Policy Management Handlers

// CreateRoleHandler creates a new role as a subject attribute with parent "role" attribute
// @Summary Create a new role
// @Description Create a new role in the RBAC system as a subject attribute
// @Tags 2-advanced-rbac
// @Accept json
// @Produce json
// @Param role body CreateRoleRequest true "Role creation request"
// @Success 201 {object} CreateRoleResponse "Role created successfully"
// @Failure 400 {object} ErrorResponse "Bad request"
// @Failure 500 {object} ErrorResponse "Internal server error"
// @Router /api/v1/rbac/roles [post]
func (h *BaseServer) CreateRoleHandler(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")

	// Parse request body
	var roleRequest CreateRoleRequest

	if err := json.NewDecoder(r.Body).Decode(&roleRequest); err != nil {
		w.WriteHeader(http.StatusBadRequest)
		json.NewEncoder(w).Encode(map[string]string{
			"error":   "invalid_request",
			"message": "Invalid JSON in request body",
		})
		return
	}

	// Validate required fields
	if roleRequest.Name == "" {
		w.WriteHeader(http.StatusBadRequest)
		json.NewEncoder(w).Encode(map[string]string{
			"error":   "validation_error",
			"message": "Role name is required",
		})
		return
	}

	// Initialize properties map if nil
	if roleRequest.Properties == nil {
		roleRequest.Properties = make(map[string]string)
	}

	// Add role-specific properties
	roleRequest.Properties["type"] = "role"
	roleRequest.Properties["parent"] = "role" // Parent attribute is "role"
	if roleRequest.Description != "" {
		roleRequest.Properties["description"] = roleRequest.Description
	}

	// Create role as a subject attribute using SubjectBuilder
	roleBuilder := h.service.CreateRole(roleRequest.Name, roleRequest.Properties)

	// Save to database
	if err := roleBuilder.Create(h.service.GetStore()); err != nil {
		h.log.Error().Err(err).Msgf("Failed to create role: %s", roleRequest.Name)
		w.WriteHeader(http.StatusInternalServerError)
		json.NewEncoder(w).Encode(map[string]string{
			"error":   "creation_failed",
			"message": "Failed to create role",
		})
		return
	}

	h.log.Info().Msgf("Created role: %s", roleRequest.Name)

	// Return success response
	w.WriteHeader(http.StatusCreated)
	response := CreateRoleResponse{
		Message: "Role created successfully",
		Role: RoleDetails{
			Name:        roleRequest.Name,
			Description: roleRequest.Description,
			Properties:  roleRequest.Properties,
			EntityID:    roleBuilder.Subject.EntityID,
		},
	}
	json.NewEncoder(w).Encode(response)
}

// UpdateRoleHandler updates an existing role
// @Summary Update a role
// @Description Update an existing role in the RBAC system
// @Tags 2-advanced-rbac
// @Accept json
// @Produce json
// @Param id path string true "Role ID"
// @Param role body UpdateRoleRequest true "Role update request"
// @Success 200 {object} UpdateRoleResponse "Role updated successfully"
// @Failure 400 {object} ErrorResponse "Bad request"
// @Failure 404 {object} ErrorResponse "Role not found"
// @Failure 500 {object} ErrorResponse "Internal server error"
// @Router /api/v1/rbac/roles/{id} [put]
func (h *BaseServer) UpdateRoleHandler(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusNotImplemented)
	json.NewEncoder(w).Encode(map[string]string{"error": "not_implemented", "message": "Update role not implemented"})
}

// GetRolesHandler retrieves all roles in the system
// @Summary List all roles
// @Description Get a list of all roles in the RBAC system
// @Tags 2-advanced-rbac
// @Produce json
// @Success 200 {array} RoleDetails "List of roles"
// @Failure 500 {object} ErrorResponse "Internal server error"
// @Router /api/v1/rbac/roles [get]
func (h *BaseServer) GetRolesHandler(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusNotImplemented)
	json.NewEncoder(w).Encode(map[string]string{"error": "not_implemented", "message": "Get roles not implemented"})
}

// DeleteRoleHandler deletes a role
// @Summary Delete a role
// @Description Delete an existing role from the RBAC system
// @Tags 2-advanced-rbac
// @Param id path string true "Role ID"
// @Success 204 "Role deleted successfully"
// @Failure 400 {object} ErrorResponse "Bad request"
// @Failure 404 {object} ErrorResponse "Role not found"
// @Failure 500 {object} ErrorResponse "Internal server error"
// @Router /api/v1/rbac/roles/{id} [delete]
func (h *BaseServer) DeleteRoleHandler(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusNotImplemented)
	json.NewEncoder(w).Encode(map[string]string{"error": "not_implemented", "message": "Delete role not implemented"})
}

// CreatePermissionHandler creates a new permission
// @Summary Create a new permission
// @Description Create a new permission in the RBAC system
// @Tags 2-advanced-rbac
// @Accept json
// @Produce json
// @Param permission body CreatePermissionRequest true "Permission creation request"
// @Success 201 {object} CreatePermissionResponse "Permission created successfully"
// @Failure 400 {object} ErrorResponse "Bad request"
// @Failure 500 {object} ErrorResponse "Internal server error"
// @Router /api/v1/rbac/permissions [post]
func (h *BaseServer) CreatePermissionHandler(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusNotImplemented)
	json.NewEncoder(w).Encode(map[string]string{"error": "not_implemented", "message": "Create permission not implemented"})
}

// GetPermissionsHandler retrieves all permissions
// @Summary List all permissions
// @Description Get a list of all permissions in the RBAC system
// @Tags 2-advanced-rbac
// @Produce json
// @Success 200 {array} PermissionDetails "List of permissions"
// @Failure 500 {object} ErrorResponse "Internal server error"
// @Router /api/v1/rbac/permissions [get]
func (h *BaseServer) GetPermissionsHandler(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusNotImplemented)
	json.NewEncoder(w).Encode(map[string]string{"error": "not_implemented", "message": "Get permissions not implemented"})
}

// AssignPermissionHandler assigns a permission to a role
// @Summary Assign permission to role
// @Description Assign an existing permission to a role in the RBAC system
// @Tags 2-advanced-rbac
// @Accept json
// @Produce json
// @Param assignment body AssignPermissionRequest true "Permission assignment request"
// @Success 200 {object} AssignPermissionResponse "Permission assigned successfully"
// @Failure 400 {object} ErrorResponse "Bad request"
// @Failure 404 {object} ErrorResponse "Role or permission not found"
// @Failure 500 {object} ErrorResponse "Internal server error"
// @Router /api/v1/rbac/permissions/assign [post]
func (h *BaseServer) AssignPermissionHandler(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusNotImplemented)
	json.NewEncoder(w).Encode(map[string]string{"error": "not_implemented", "message": "Assign permission to role not implemented"})
}

// RemovePermissionHandler removes a permission from a role
// @Summary Remove permission from role
// @Description Remove an assigned permission from a role in the RBAC system
// @Tags 2-advanced-rbac
// @Accept json
// @Produce json
// @Param removal body RemovePermissionRequest true "Permission removal request"
// @Success 200 {object} RemovePermissionResponse "Permission removed successfully"
// @Failure 400 {object} ErrorResponse "Bad request"
// @Failure 404 {object} ErrorResponse "Role or permission not found"
// @Failure 500 {object} ErrorResponse "Internal server error"
// @Router /api/v1/rbac/permissions/remove [post]
func (h *BaseServer) RemovePermissionHandler(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusNotImplemented)
	json.NewEncoder(w).Encode(map[string]string{"error": "not_implemented", "message": "Remove permission from role not implemented"})
}