# pkg/server/rebac_handlers.go
package server

import (
	"encoding/json"
	"net/http"
)

// ReBAC Schema Management Handlers

// CreateReBACSchemaHandler creates a new ReBAC schema
// @Summary Create a new ReBAC schema
// @Description Create a new Relationship-Based Access Control schema
// @Tags 4-advanced-rebac
// @Accept json
// @Produce json
// @Param schema body CreateReBACSchemaRequest true "ReBAC schema creation request"
// @Success 201 {object} CreateReBACSchemaResponse "ReBAC schema created successfully"
// @Failure 400 {object} ErrorResponse "Bad request"
// @Failure 500 {object} ErrorResponse "Internal server error"
// @Router /api/v1/rebac/schemas [post]
func (h *BaseServer) CreateReBACSchemaHandler(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusNotImplemented)
	json.NewEncoder(w).Encode(map[string]string{"error": "not_implemented", "message": "Create ReBAC schema not implemented"})
}

// GetReBACSchemasHandler retrieves all ReBAC schemas
// @Summary List all ReBAC schemas
// @Description Get a list of all Relationship-Based Access Control schemas
// @Tags 4-advanced-rebac
// @Produce json
// @Success 200 {array} ReBACSchemaDetails "List of ReBAC schemas"
// @Failure 500 {object} ErrorResponse "Internal server error"
// @Router /api/v1/rebac/schemas [get]
func (h *BaseServer) GetReBACSchemasHandler(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusNotImplemented)
	json.NewEncoder(w).Encode(map[string]string{"error": "not_implemented", "message": "Get ReBAC schemas not implemented"})
}

// UpdateReBACSchemHandler updates an existing ReBAC schema
// @Summary Update a ReBAC schema
// @Description Update an existing Relationship-Based Access Control schema
// @Tags 4-advanced-rebac
// @Accept json
// @Produce json
// @Param id path string true "Schema ID"
// @Param schema body UpdateReBACSchemaRequest true "ReBAC schema update request"
// @Success 200 {object} UpdateReBACSchemaResponse "ReBAC schema updated successfully"
// @Failure 400 {object} ErrorResponse "Bad request"
// @Failure 404 {object} ErrorResponse "Schema not found"
// @Failure 500 {object} ErrorResponse "Internal server error"
// @Router /api/v1/rebac/schemas/{id} [put]
func (h *BaseServer) UpdateReBACSchemHandler(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusNotImplemented)
	json.NewEncoder(w).Encode(map[string]string{"error": "not_implemented", "message": "Update ReBAC schema not implemented"})
}

// CreateRelationTypeHandler creates a new relation type
// @Summary Create a new relation type
// @Description Create a new relation type for ReBAC schemas
// @Tags 4-advanced-rebac
// @Accept json
// @Produce json
// @Param relationType body CreateRelationTypeRequest true "Relation type creation request"
// @Success 201 {object} CreateRelationTypeResponse "Relation type created successfully"
// @Failure 400 {object} ErrorResponse "Bad request"
// @Failure 500 {object} ErrorResponse "Internal server error"
// @Router /api/v1/rebac/relation-types [post]
func (h *BaseServer) CreateRelationTypeHandler(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusNotImplemented)
	json.NewEncoder(w).Encode(map[string]string{"error": "not_implemented", "message": "Create relation type not implemented"})
}

// GetRelationTypesHandler retrieves all relation types
// @Summary List all relation types
// @Description Get a list of all relation types used in ReBAC schemas
// @Tags 4-advanced-rebac
// @Produce json
// @Success 200 {array} RelationTypeDetails "List of relation types"
// @Failure 500 {object} ErrorResponse "Internal server error"
// @Router /api/v1/rebac/relation-types [get]
func (h *BaseServer) GetRelationTypesHandler(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusNotImplemented)
	json.NewEncoder(w).Encode(map[string]string{"error": "not_implemented", "message": "Get relation types not implemented"})
}